+++ fragments/labels/mattermost.sh
@@ -1,12 +1,9 @@
 mattermost)
     name="Mattermost"
     type="dmg"
-    if [[ $(arch) == i386 ]]; then
-      archiveName="mac-x64.dmg"
-    elif [[ $(arch) == arm64 ]]; then
-      archiveName="mac-m1.dmg"
-    fi
+    archiveName="mac-universal.dmg"
     downloadURL=$(downloadURLFromGit mattermost desktop)
     appNewVersion=$(versionFromGit mattermost desktop )
     expectedTeamID="UQ8HT4Q2XM"
+    Mattermost Helper (Renderer).app app.asar
     ;;